style: 자바 스타일 적용
