+++ src/test/java/acceptance/OrderAcceptanceTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
+import common.AcceptanceTest;
 import io.restassured.RestAssured;
 import java.util.List;
 import java.util.Map;
