+++ src/test/java/common/AcceptanceTest.java
@@ -1,0 +1,17 @@
+package common;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.boot.web.server.LocalServerPort;
+
+@IntegrationTest
+public abstract class AcceptanceTest {
+
+    @LocalServerPort
+    private int port;
+
+    @BeforeEach
+    void setUp() {
+        RestAssured.port = port;
+    }
+}
