+++ src/test/java/common/IntegrationTest.java
@@ -1,6 +1,7 @@
-package kitchenpos.common;
+package common;
 
 import java.lang.annotation.ElementType;
+import java.lang.annotation.Inherited;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
@@ -16,5 +17,6 @@ import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
         classes = Application.class
 )
 @ExtendWith(DataBaseCleaner.class)
+@Inherited
 public @interface IntegrationTest {
 }
