+++ src/test/java/kitchenpos/ui/TableRestControllerTest.java
@@ -3,10 +3,10 @@ package kitchenpos.ui;
 import static fixture.MenuFixtures.후라이드치킨_메뉴;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import common.IntegrationTest;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.common.IntegrationTest;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderTable;
